+++ tests/test_encryption.py
@@ -1,0 +1,12 @@
+from prefab_cloud_python.encryption import Encryption
+
+
+class TestEncryption:
+    def test_encryption(self):
+        secret = Encryption.generate_new_hex_key()
+
+        enc = Encryption(secret)
+
+        clear_text = "hello world"
+        encrypted = enc.encrypt(clear_text)
+        assert enc.decrypt(encrypted) == clear_text